# optimized performance for catalog render and map calculation