related issue #1426
For 1 million data points

> The render time (each 100K points) is improved from 600ms to 250ms.

> The time for calculating catalog map is improved from 280ms to 25ms.